feat: convert inventory to class
